% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coordenadas}
\alias{coordenadas}
\title{Ocurrence points for four species in the Brazilian Atlantic Forest.}
\format{A dataframe with 618 rows and 3 columns:
\describe{
  \item{sp}{Species name}
  \item{lat}{Latitude}
  \item{lon}{Longitude}
}}
\usage{
coordenadas
}
\description{
A dataframe with 618 occurrence points from four species
}
\keyword{datasets}
