% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffer.R
\name{create_buffer}
\alias{create_buffer}
\title{Samples pseudoabsences inside a geographic buffer}
\usage{
create_buffer(coord, n.back, buffer.type, seed = 512, predictors)
}
\arguments{
\item{coord}{A data frame with occurrence data. It should contain only two columns:
lon and lat, in that order.}

\item{n.back}{Number of pseudoabsence points}

\item{buffer.type}{Character string indicating whether the buffer should be
calculated using the mean, median or maximum distance between occurrence points}

\item{seed}{for reproducibility purposes}

\item{predictors}{A RasterStack of predictor variables}
}
\value{
Table of pseudoabsence points sampled within the selected distance
}
\description{
Samples pseudoabsences inside a geographic buffer
}
\details{
The sampling is performed by dismo::randomPoints() excluding the presence points (exclupep =TRUE)
}
\examples{
create_buffer(coordinates = coordenadas, n.back = 500, buffer.type = "mean",
 predictors = variaveis_preditoras)

}
\references{
VanDerWal, J., Shoo, L. P., Graham, C., & Williams, S. E. (2009). Selecting pseudo-absence data for presence-only distribution modeling: How far should you stray from what you know? Ecological Modelling, 220(4), 589-594. doi:10.1016/j.ecolmodel.2008.11.010
}
\seealso{
\code{\link[raster]{buffer}}

\code{\link[dismo]{randomPoints}}
}
\author{
Felipe Barros

Fabrício Vilasboas

Andrea Sánchez-Tapia
}
