% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_model.R
\name{crop_model}
\alias{crop_model}
\title{Crop environmental raster based on speficif buffer}
\usage{
crop_model(modelo, mask)
}
\arguments{
\item{modelo}{raster with the ecological niche model}

\item{mask}{shapefile object to be used as mask for cropping the model}
}
\value{
Croppped environmental raster based on specific buffer
}
\description{
Crop environmental raster based on speficif buffer
}
\seealso{
\code{\link[raster]{crop}}

\code{\link[dismo]{mask}}
}
\author{
Andrea Sánchez-Tapia
}
