% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_model.R
\name{do_any}
\alias{do_any}
\title{Fits ecological niche models using any dismo algorithm.}
\usage{
do_any(species_name, algo = c("bioclim", "maxent", "mahal", "domain"),
  coordinates, partitions, buffer = FALSE, seed = 512, predictors,
  models.dir = "./models", project.model = FALSE, projections = NULL,
  mask = NULL, write_png = FALSE, n.back, bootstrap = T,
  boot_proportion = 0.8, n_boot = 10, crossvalidation = F, n_cv = 10)
}
\arguments{
\item{species_name}{A character string with the species name}

\item{algo}{The algorithm to be fitted \code{c("bioclim", "maxent","domain",
"mahal", "glm", "svm.k","svm.e","rf")}}

\item{coordinates}{A two-column data frame with the occurrence points}

\item{partitions}{The number of partitions for a cross validation}

\item{buffer}{Defines if a buffer will be used to sample pseudo-absences
(F, "mean", "median", "max")}

\item{seed}{For reproducibility purposes}

\item{predictors}{A RasterStack of predictor variables}

\item{models.dir}{Folder path to save the output files}

\item{project.model}{Logical, whether to perform a projection}

\item{projections}{The RasterStack of projeciton variables}

\item{mask}{A SpatialPolygonsDataFrame to be used to mask the final models}

\item{write_png}{Logical, whether png files will be written}

\item{n.back}{Number of pseudoabsence points}
}
\value{
A data frame with the evaluation statistics (TSS, AUC, etc.)
}
\description{
Fits ecological niche models using any dismo algorithm.
}
