% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_enm.R
\name{do_enm}
\alias{do_enm}
\title{Fits ecological niche models for various algorithms.}
\usage{
do_enm(especie, coordinates, partitions, buffer, seed, predictors, models.dir,
  project.model, projections, mask, n.back, write_png, maxent = T,
  domain = T, bioclim = T, rf = T, glm = T, svm = T, svm2 = T,
  mahal = T)
}
\arguments{
\item{coordinates}{A three-column data frame with the occurrence points of all species (sp, lat, lon)}

\item{buffer}{Defines if a buffer will be used to sample pseudo-absences (F, "mean", "median", "max")}

\item{seed}{For reproducibility purposes}

\item{predictors}{A RasterStack of predictor variables}

\item{models.dir}{Folder path to save the output files}

\item{mask}{A SpatialPolygonsDataFrame to be used as a mask to cut the final models}

\item{n.back}{Number of pseudoabsence points}

\item{write_png}{Logical, whether png files will be written, defaults to F}

\item{sp}{A character string with the species name}
}
\value{
A data frame with the evaluation statistics (TSS, AUC, and their respective thresholds)
}
\description{
Fits ecological niche models for various algorithms.
}
