% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahal.R
\name{do_mahal}
\alias{do_mahal}
\title{Fits ecological niche models using Mahalanobis distance.}
\usage{
do_mahal(sp, coordinates, partitions, buffer = FALSE, seed = 512,
  predictors, models.dir = "./models", project.model = FALSE,
  projections = NULL, mask = NULL, write_png = FALSE, n.back)
}
\arguments{
\item{sp}{A character string with the species name}

\item{coordinates}{A two-column data frame with the occurrence points}

\item{partitions}{The number of partitions for a cross validation}

\item{buffer}{Defines if a buffer will be used to sample pseudo-absences
(F, "mean", "median", "max")}

\item{seed}{For reproducibility purposes}

\item{predictors}{A RasterStack of predictor variables}

\item{models.dir}{Folder path to save the output files}

\item{project.model}{Logical, whether to perform a projection}

\item{projections}{The RasterStack of projeciton variables}

\item{mask}{A SpatialPolygonsDataFrame to be used to mask the final models}

\item{write_png}{Logical, whether png files will be written}

\item{n.back}{Number of pseudoabsence points}
}
\value{
A data frame with the evaluation statistics (TSS, AUC, etc.)
}
\description{
Fits ecological niche models using Mahalanobis distance.
}
