% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_many.R
\name{do_many}
\alias{do_many}
\title{Fits ecological niche models for various algorithms.}
\usage{
do_many(species_name, bioclim = FALSE, domain = FALSE, glm = FALSE,
  mahal = FALSE, maxent = FALSE, maxnet = FALSE, rf = FALSE,
  svmk = FALSE, svme = FALSE, mindist = FALSE, centroid = FALSE,
  brt = FALSE, ...)
}
\arguments{
\item{species_name}{A character string with the species name. Because species name will be used as 
a directory name, avoid non-ASCII characters, spaces and ponctuation marks. 
Recommendation is to adopt "Genus_species" format. See names in \code{\link{coordenadas}} as an example.}

\item{bioclim}{Execute bioclim from the dismo implementation}

\item{domain}{Execute domain from the dismo implementation}

\item{glm}{Execute GLM as suggested by the dismo documentation}

\item{mahal}{Execute mahalanobis distance from the dismo implementation}

\item{maxent}{Execute maxent from the dismo implementation}

\item{maxnet}{Execute maxent from the maxnet implementation}

\item{rf}{Execute random forests from randomForest() as suggested by the dismo documentation}

\item{svmk}{Execute svm from kernlab package}

\item{svme}{Execute svm from e1071 package}

\item{mindist}{Execute minimum euclidean distance}

\item{centroid}{Execute euclidean distance to the environmental centroid}

\item{brt}{Execute boosted regression trees}

\item{...}{Any parameter from \link{do_any}}
}
\value{
A set of ecological niche models for each partition and algorithm,
        written in the \code{models_dir} subfolder
}
\description{
Fits ecological niche models for various algorithms.
}
\author{
Andrea Sánchez-Tapia
}
