% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxent.R
\name{do_maxent}
\alias{do_maxent}
\title{Faz modelagem de distribuição de espécies com algoritmo Maxent}
\usage{
do_maxent(sp, coordinates, partitions, buffer = FALSE, seed = 512,
  predictors, models.dir, project.model, projections, mask, n.back = 500)
}
\arguments{
\item{sp}{Um nome de espécie}

\item{coordinates}{Uma tabela com pontos de ocorrência}

\item{buffer}{Define se será usado buffer e de que tipo ("mean" ou "max")}

\item{seed}{Para reprodutibilidade}

\item{predictors}{Objeto do tipo RasterStack com variáveis preditoras}

\item{models.dir}{Path do diretório onde serão escritos os arquivos de saída}

\item{mask}{Objeto do tipo SpatialPolygonsDataFrame com máscara}

\item{n.back}{Número de pontos de background}
}
\value{
Um data.frame com metadados da modelagem (TSS, AUC, algoritmo etc.)
}
\description{
Faz modelagem de distribuição de espécies com algoritmo Maxent
}
