% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{do_SVM}
\alias{do_SVM}
\title{Fits ecological niche models using SVM from package kernlab.}
\usage{
do_SVM(sp, coordinates, partitions, buffer, seed, predictors, models.dir,
  project.model, projections, mask, write_png, n.back)
}
\arguments{
\item{sp}{A character string with the species name}

\item{coordinates}{A two-column data frame with the occurrence points}

\item{partitions}{The number of partitions for a cross validation}

\item{buffer}{Defines if a buffer will be used to sample pseudo-absences
(F, "mean", "median", "max")}

\item{seed}{For reproducibility purposes}

\item{predictors}{A RasterStack of predictor variables}

\item{models.dir}{Folder path to save the output files}

\item{project.model}{Logical, whether to perform a projection}

\item{projections}{The RasterStack of projeciton variables}

\item{mask}{A SpatialPolygonsDataFrame to be used to mask the final models}

\item{write_png}{Logical, whether png files will be written}

\item{n.back}{Number of pseudoabsence points}
}
\value{
A data frame with the evaluation statistics (TSS, AUC, and their respective thresholds)
}
\description{
Fits ecological niche models using SVM from package kernlab.
}
