% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_model}
\alias{ensemble_model}
\title{Create ensemble models from several algorithms.}
\usage{
ensemble_model(species_name, occurrences, lon = "lon", lat = "lat",
  models_dir = "./models", final_dir = "final_models",
  ensemble_dir = "ensemble", proj_dir = "present",
  which_models = c("raw_mean"), consensus = FALSE,
  consensus_level = 0.5, write_png = T, scale_models = TRUE,
  write_raw_map = F)
}
\arguments{
\item{species_name}{A character string with the species name}

\item{occurrences}{A two-column data frame with the occurrence points}

\item{lon}{the name of the longitude column. defaults to "lon"}

\item{lat}{the name of the latitude column. defaults to "lat"}

\item{models_dir}{Character string. Folder path where the input files
are located}

\item{final_dir}{Character string, name of the subfolder where the files
for the final models are located}

\item{ensemble_dir}{Character string, name of the folder to save the output
files. A subfolder will be created.}

\item{proj_dir}{Character. The name of the subfolder with the projection.
Defaults to "present" but can be set according to the other projections (i.e.
to execute the function in projected models)}

\item{which_models}{Which final_model() will be used? Currently it can be:
\describe{
  \item{\code{weighted_AUC} or \code{weighted_TSS}}{the models weighted
  by TSS or AUC}
  \item{\code{raw_mean}}{the mean of the selected raw models}
  \item{\code{bin_mean_th}}{the binary model created by cutting \code{raw_mean} by the mean of the thresholds that
   maximize the selected evaluation metric (e.g. TSS (\code{spec_sens}) or other dismo thresholds)}
   \item{\code{cut_mean_th}}{the cut model created by recovering \code{raw_mean} values above the mean threshold that
   maximizes the selected evaluation metric (e.g. TSS (\code{spec_sens}) or other dismo thresholds)}
  \item{\code{bin_mean}}{the mean of the selected binary models}
  \item{\code{bin_consensus}}{the binary consensus from \code{bin_mean}}.
  \item{\code{cut_mean}}{the mean of the selected cut models}
}}

\item{consensus}{Logical. Will a consensus between the algorithms be applied?}

\item{consensus_level}{Threshold for the consensus rule, betwen 0 and 1
(0.5 means a majority rule).}

\item{write_png}{Write png? Defaults to TRUE}

\item{scale_models}{Logical. If TRUE (default), sets the values of the input models between 0 and 1.}

\item{write_raw_map}{Create a mean raw map without margins}
}
\value{
A rasterStack with the minimun, maximum, median, mean and standard
deviation of the assembled models. A set of ensemble models and figures
(optional) written in the \code{ensemble_dir} subfolder
}
\description{
\code{ensemble} takes the input of \code{final_model} and builds an ensemble
 model by calculating the mean of each final model per algorithm.
}
\seealso{
\link{final_model}
}
