% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{This function calculates the mean distance to the centroid of a distribution}
\usage{
euclidean(predictors, occurrences, algo = "centroid", filename = "",
  ...)
}
\arguments{
\item{predictors}{is an environmental variables RasterStack}

\item{occurrences}{are the occurrence points}

\item{algo}{is "minimum" or "centroid"}

\item{filename}{Optional. The raster that will be created in disk}

\item{...}{other parameters in raster::writeRaster()}
}
\description{
This function calculates the mean distance to the centroid of a distribution
}
\examples{
centr <- euclidean(example_vars, occurrences = coordenadas[[1]][,c(2,3)])
raster::plot(centr)
}
