% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_mask}
\alias{example_mask}
\title{Example mask for cropping the partition models}
\format{
A SpatialPolygonsDataFrame of the Brazilian Atlantic Forest, based
on IBGE
}
\usage{
example_mask
}
\description{
Example mask for cropping the partition models
}
\keyword{datasets}
