% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_vars}
\alias{example_vars}
\title{Predictor variables}
\format{A RasterStack with 6 predictor variables issued from a PCA, at 10min resolution}
\usage{
example_vars
}
\description{
Predictor variables
}
\keyword{datasets}
