% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_model.R
\name{finalModel}
\alias{finalModel}
\title{Joins ENM from several partitions, creating a model per algorithm.}
\usage{
finalModel(sp, select.partitions = T, algoritmos = c("maxent", "rf", "svm"),
  threshold = c("spec_sens"), TSS.value = 0.7, models.dir = "./models",
  final.dir = "final_models")
}
\arguments{
\item{sp}{A character string with the species name}

\item{select.partitions}{TRUE ou FALSE}

\item{algoritmos}{Which algorithms will be processed. Defaults to c("maxent", "rf", "svm")}

\item{threshold}{Which cutting threshold will be used, defaults to "spec_sens"}

\item{TSS.value}{Threshold to select models from TSS values (between -1 and 1)}

\item{models.dir}{Character string. Folder path where the input files are located}

\item{final.dir}{Character string. Name of the folder to save the output files. A subfolder will be created.}
}
\value{
A character string with the initial and final execution times.
}
\description{
Joins ENM from several partitions, creating a model per algorithm.
}
