% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_any.R, R/do_many.R, R/model_fit.R
\name{do_any}
\alias{do_any}
\alias{do_many}
\alias{fit}
\title{Model fitting, predicting and evaluating of ecological niche models using several algorithms}
\usage{
do_any(species_name, predictors, models_dir = "./models",
  algo = c("bioclim"), project_model = FALSE,
  proj_data_folder = "./data/proj", mask = NULL, write_png = FALSE,
  write_bin_cut = FALSE, threshold = "spec_sens", conf_mat = TRUE,
  equalize = TRUE, proc_threshold = 0.5, ...)

do_many(species_name, bioclim = FALSE, domain = FALSE, glm = FALSE,
  mahal = FALSE, maxent = FALSE, maxnet = FALSE, rf = FALSE,
  svmk = FALSE, svme = FALSE, brt = FALSE, ...)
}
\arguments{
\item{species_name}{A character string with the species name. Because species name will be used as
a directory name, avoid non-ASCII characters, spaces and punctuation marks.
Recommendation is to adopt "Genus_species" format. See names in \code{\link{coordenadas}} as an example.}

\item{predictors}{An object with the rasters to be cut.
Accepts an object of type RasterStack, generated by the function \code{\link[raster]{stack}}.}

\item{models_dir}{Folder path to save the output files.}

\item{algo}{The algorithm to be fitted \code{c("bioclim", "brt", "domain",
"glm", "maxent", "mahal",
"svme", "svmk", "rf")}.}

\item{project_model}{Logical, whether to perform a projection.}

\item{proj_data_folder}{Path to directory with projections containing one or more folders with the projection datasets (e.g. "./env/proj/proj1"). Projection diretctory should only contain raster files corresponding to the environmental variables. If more than one projection, each projection should be at one directory (e.g. "./env/proj/proj1" and "./env/proj/proj2") and equivalent raster files at diferent subdirectories must have the same names (e.g. "./env/proj/proj1/layer1.asc" and "./env/proj/proj2/layer1.asc").}

\item{mask}{A SpatialPolygonsDataFrame to be used to mask the final models.}

\item{write_png}{Logical, whether png files will be written.}

\item{write_bin_cut}{Logical, whether binary and cut model files(.tif, .png) should be written.}

\item{threshold}{Character string indicating threshold (cut-off) to transform model predictions
to a binary score.
as in \code{\link[dismo]{threshold}}: "kappa", "spec_sens", "no_omission", "prevalence",
"equal_sens_spec", "sensitivity". Default value is "spec_sens".}

\item{conf_mat}{Logical, whether confusion tables should be written in the HD.}

\item{equalize}{Logical, whether the number of presences and absences should be
equalized in randomForest and brt.}

\item{proc_threshold}{Numeric, value from 0 to 100 that will be used as (E) for partialROC calculations in \code{\link[kuenm]{kuenm_proc}}. Default is \code{proc_threshold = 5}.}

\item{...}{Any parameter from \link{do_any}}

\item{bioclim}{Execute bioclim algorithm from the \pkg{dismo} implementation
with \code{\link[dismo]{bioclim}} function.}

\item{domain}{Execute domain from the \pkg{dismo} implementation with
\code{\link[dismo]{domain}} function.}

\item{glm}{Execute GLM as suggested by the \pkg{dismo} documentation with
\code{\link{glm}} and \code{\link{step}}.}

\item{mahal}{Execute Mahalanobis distance from the \pkg{dismo} implementation
with \code{\link[dismo]{mahal}}.}

\item{maxent}{Execute Maxent algorithm from the \pkg{dismo} implementation with
\code{\link[dismo]{maxent}} function.}

\item{maxnet}{Execute Maxent algorithm from the \pkg{maxnet} implementation with
\code{\link[maxnet]{maxnet}} function.}

\item{rf}{Execute Random forests algorithm from \pkg{randomForest} package with
function\code{\link[randomForest]{tuneRF}} as suggested by the \pkg{dismo} documentation.}

\item{svmk}{Execute Support Vector Machines (SVM) algorithm from
\pkg{kernlab} package with \code{\link[kernlab]{ksvm}} function.}

\item{svme}{Execute Support Vector Machines (SVM) algorithm from \pkg{e1071}
package with \code{\link[e1071]{best.tune}} function.}

\item{brt}{Execute Boosted Regression Trees with
\code{\link[dismo]{gbm.step}} from \pkg{dismo}.}

\item{...}{Other arguments from \code{\link[kuenm]{kuenm_proc}}.}
}
\value{
Writes on disk model for each partition, a .csv file with evaluation statistics (TSS, AUC, etc).
}
\description{
\code{do_any} reads the output from \code{\link{setup_sdmdata}} and computes ecological niche models for a species based on an algorithm specified by the user. It fits the model, calculates the predicted values and basic statistics for model evaluation. in addition to commonly adopted metrics such as AUC and TSS, this package also calculates partial ROC (pROC) \insertCite{for details on model evaluation see @phillips_maximum_2006, @peterson_ecological_2011}{modleR}.
\code{do_any} performs one algorithm at a time. \code{do_many} runs internally \code{\link{do_any}} and can be used to run multiple algorithms at a time. See \strong{Details} in \code{\link{do_any}} for a description of how each algorithm is implemented. Given that there are "\emph{no silver bullets in correlative ecological niche modeling}" \insertCite{qiao_no_2015}{modleR} the choice of which algorithm to run is on the user. See details for a description of all algorithms supported in this package.
}
\details{
See below for a description on the implementation of the algorithms supported in this package.
\describe{
\item{Bioclim}{
Specified by \code{algo = "bioclim"} uses \code{\link[dismo]{bioclim}} function in \pkg{dismo} package \insertCite{hijmans_dismo_2017}{modleR}. Bioclim is the  climate-envelope-model implemented by Henry Nix \insertCite{nix_biogeographic_1986}{modleR}, the first species distribution modelling package. It is based on climate interpolation methods and despite its limitations it is still used in ecological niche modeling, specially for exploration and teaching purposes \insertCite{@see also @booth_bioclim_2014}{modleR}. In this package it is implemented by the function \code{\link[dismo]{bioclim}}, evaluated and predicted using \code{\link[dismo]{evaluate}} and \code{\link[dismo]{predict}} also from \pkg{dismo} package.
}
\item{Boosted Regression Trees (BRT)}{
Specified by \code{algo="brt"} uses \code{\link[dismo]{gbm.step}} function from \pkg{dismo} package. Runs the cross-validation procedure of \insertCite{hastie_elements_2001;textual}{modleR} \insertCite{@see also @elith_working_2009}{modleR}. It consists in a regression modeling technique combined with the boosting method, a method for combining many simple models. It is implemented by the function \code{\link[dismo]{gbm.step}} as a regression with the response variable set to bernoulli distribution, evaluated and predicted using \code{\link[dismo]{evaluate}} and \code{\link[dismo]{predict}} from \pkg{dismo} package.
}
\item{Domain}{
Specified by \code{algo="domain"} uses \code{\link[dismo]{domain}} function from \pkg{dismo} package. Computes point-to-point similarity based on Gower distance between environmental variables \insertCite{carpenter_domain_1993}{modleR}. \insertCite{hijmans_dismo_2017}{modleR} state that one should use it with caution because it does not perform well compared to other algorithms \insertCite{elith_novel_2006,hijmans_ability_2006}{modleR}. We add that it is a slow algorithm. In this package it is implemented by the function \code{\link[dismo]{domain}}, evaluated and predicted using \code{\link[dismo]{evaluate}} and \code{\link[dismo]{predict}} also from \pkg{dismo} package.
}
\item{Generalized Linear Model (GLM)}{
Specified by \code{algo="glm"} runs a GLM with modeling presence and absences as a response variable following a binomial error distribution. It runs runs a step-wise model selection based on AIC both backward and forward considering all possible combinations of predictor variables in the rasterStack. In this package it is implemented using functions \code{glm} and \code{step} to fit a model and choose a model by AIC in a stepwise procedure. Model is evaluated and predicted using \code{\link[dismo]{evaluate}} function from \pkg{dismo} and \code{\link[raster]{predict}} function from \pkg{raster} package both with argument \code{type="response"} to return values in the scale of the response variable.
}
\item{Mahalanobis}{
Specified by \code{algo="mahal"} uses \code{\link[dismo]{mahal}} function from \pkg{dismo} package. Corresponds to a distribution model based on Mahalanobis distance, a measure of the distance between a point P and a distribution D \insertCite{mahalanobis_generalized_1936}{modleR}. In this package it is implemented by the function \code{\link[dismo]{mahal}}, evaluated and predicted using \code{\link[dismo]{evaluate}} and \code{\link[dismo]{predict}} also from \pkg{dismo} package.
}
\item{Maximum Entropy (Maxent)}{
Specified either by \code{algo="maxent"} or \code{algo="maxnet"} corresponding to implementation by \pkg{dismo} \insertCite{hijmans_dismo_2017}{modleR} and \pkg{maxnet} \insertCite{phillips_maxnet_2017}{modleR} packages respectivelly. Maxent is a machine learning method for modeling species distributions based in incomplete data allowing ENM with presence-only data \insertCite{phillips_maximum_2006}{modleR}. If \code{algo="maxent"} model is fitted by the function \code{\link[dismo]{maxent}}, evaluated and predicted using  \code{\link[dismo]{evaluate}} and \code{\link[dismo]{predict}} also in \pkg{dismo} package. If \code{algo="maxnet"} model is fitted by the function \code{\link[maxnet]{maxnet}} from \pkg{maxnet} package, evaluated using \code{\link[dismo]{evaluate}} from \pkg{dismo} package with argument \code{type="logistic"} and predicted using \code{\link[raster]{predict}} function from \pkg{raster} package.
}
\item{Random Forest}{
Specified by \code{algo="rf"} uses \code{\link[randomForest]{tuneRF}} function from \pkg{ramdomForest} package \insertCite{liaw_classification_2002}{modleR}. Corresponds to machine learning regression based on decision trees. In this package uses \code{\link[randomForest]{tuneRF}} function with the optimal number of variables available for splitting at each tree node (i.e. mtry) found as set by parameter \code{doBest = TRUE}. Random Forest model is evaluated with \code{\link[dismo]{evaluate}} function from \pkg{dismo} and predicted with \code{\link[raster]{predict}} function from \pkg{raster} package.
}
\item{Support Vector Machines (SVM)}{
Specified either by \code{algo = "svme"} or \code{algo = "svmk"} corresponding to implementation on \pkg{e1071} \insertCite{meyer_e1071_2017}{modleR} and \pkg{kernlab} \insertCite{karatzoglou_kernlab_2004}{modleR} packages respectivelly. SVM are supervised learning models that use learning algorithms for classification and regression analysis. In \pkg{e1071} package SVM is implemented through function \code{\link[e1071]{best.tune}} with method set to \code{"svm"} which uses RBF-kernel (radial basis function kernel) for classification. In \pkg{kernlab} package SVM is implemented through function \code{\link[kernlab]{ksvm}} also with RBF-kernel method (in this case the default method \code{"kbfdot"}). We expect both implementations to differ only in performance. Both \code{svme} and \code{svmk} are evaluated with \code{\link[dismo]{evaluate}} function from dismo and predicted with \code{\link[raster]{predict}} function from \pkg{raster} package.
}
}
}
\examples{
# run setup_sdmdata first from one species in coordenadas data
sp <- names(coordenadas)[1]
sp_coord <- coordenadas[[1]]
sp_setup <- setup_sdmdata(species_name = sp, occurrences = sp_coord, example_vars)

# run bioclim algorithm for one species
do_any(species_name = sp,
       predictors = example_vars,
       algo = "bioclim")

# run setup_sdmdata
sp <- names(coordenadas)[1]
sp_coord <- coordenadas[[1]]
sp_setup <- setup_sdmdata(species_name = sp, occurrences = sp_coord, example_vars)

# run do_many
sp_many <- do_many(species_name = sp,
                     predictors = example_vars,
                     bioclim = TRUE,
                     maxent = TRUE)
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[dismo]{bioclim}}

\code{\link[dismo]{domain}}

\code{\link{do_any}}

\code{\link[dismo]{evaluate}}

\code{\link[dismo]{maxent}}

\code{\link[maxnet]{maxnet}}

\code{\link[dismo]{mahal}}

\code{\link[dismo]{predict}} in \pkg{dismo} package

\code{\link[raster]{predict}} in \pkg{raster} package
}
