% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mascara}
\alias{mascara}
\title{Máscara para a Mata Atlântica}
\format{Um SpatialPolygonsDataFrame da Mata Atlântica}
\usage{
mascara
}
\description{
Máscara para a Mata Atlântica
}
\keyword{datasets}
