% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelr.R
\docType{package}
\name{modelr}
\alias{modelr}
\alias{modelr-package}
\title{modelr: A workflow to perform Environmental Niche Modeling based on dismo.}
\description{
The modelr package wraps commonly used enm functions into a four-step
workflow: setup_sdmdata, do_(m)any, final_models and ensemble.
}
\section{modelr functions}{

The three-step workflow implemented here ...
}

