% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modleR.R
\docType{package}
\name{modleR}
\alias{modleR}
\alias{modleR-package}
\title{modleR: A workflow to perform Environmental Niche Modeling based on dismo.}
\description{
The modleR package wraps commonly used enm functions into a four-step
workflow: setup_sdmdata, do_(m)any, final_models and ensemble.
}
\section{modleR functions}{

The three-step workflow implemented here ...
}

