% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale.layer}
\alias{rescale.layer}
\title{Rescaling rasters to values between 0 and 1.}
\usage{
rescale.layer(layers)
}
\arguments{
\item{layers}{A RasterStack or Rasterlayer of predictor variables}

\item{models_dir}{Character. Folder path where the input files are located}

\item{final_dir}{Character. Name of the folder to save the output files.
A subfolder will be created.}
}
\value{
A RasterStack or Rasterlayer with values between 0 and 1.
}
\description{
This function rescales rasters to values between 0 and 1.
}
\examples{
rescale.layer(example_vars)
}
\author{
Diogo S. B. Rocha
}
