% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_layer.R
\name{rescale_layer}
\alias{rescale_layer}
\title{Rescales rasters to values between 0 and 1}
\usage{
rescale_layer(layers)
}
\arguments{
\item{layers}{A RasterStack or RasterLayer of predictor variables to scale}
}
\value{
A RasterStack or RasterLayer with values between 0 and 1.
}
\description{
This function rescales rasters to values between 0 and 1.
}
\examples{
rescale_layer(example_vars)
}
