% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sdmdata.R
\name{setup_sdmdata}
\alias{setup_sdmdata}
\title{Prepares the dataset to perform ENM}
\usage{
setup_sdmdata(species.name = species.name, coordinates = coordinates,
  buffer = FALSE, seed = 512, predictors = predictors,
  models.dir = models.dir, plot_sdmdata = T, n.back = 1000,
  bootstrap = F, boot_proportion = 0.8, n_boot = 10,
  crossvalidation = F, partitions = partitions, n_cv = 10)
}
\arguments{
\item{species.name}{A species name}

\item{coordinates}{Uma tabela com pontos de ocorrência}

\item{buffer}{Define se será usado buffer e de que tipo ("mean" ou "max")}

\item{seed}{Para reprodutibilidade}

\item{predictors}{Objeto do tipo RasterStack com variáveis preditoras}

\item{models.dir}{Path do diretório onde serão escritos os arquivos de saída}

\item{plot_sdmdata}{Se o png vai ser criado - defaults to F}

\item{n.back}{Número de pontos de background}
}
\value{
Um data.frame com metadados da modelagem (TSS, AUC, algoritmo etc.)
}
\description{
Prepares the dataset to perform ENM
}
