% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variaveis_preditoras}
\alias{variaveis_preditoras}
\title{Variáveis preditoras}
\format{Um RasterStack com variáveis preditoras de exemplo}
\usage{
variaveis_preditoras
}
\description{
Variáveis preditoras
}
\keyword{datasets}
