% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Cleans the occurrence records}
\usage{
clean(occurrences, lon = "lon", lat = "lat", predictors,
  clean_dupl = FALSE, clean_nas = FALSE, clean_uni = FALSE)
}
\arguments{
\item{occurrences}{A data frame with occurrence data. Data must have at least columns with latitude an longitude values of species occurrences.
See \code{coordenadas} as an example.}

\item{lon}{The name of the longitude column. Defaults to "lon".}

\item{lat}{The name of the latitude column. Defaults to "lat".}

\item{predictors}{An object with the rasters to be cut.
Accepts an object of type RasterStack, generated by the function \code{\link[raster]{stack}}.}

\item{clean_dupl}{logical. If TRUE removes points with the same longitude and latitude}

\item{clean_nas}{logical. If TRUE removes points that are outside the bounds of the raster}

\item{clean_uni}{logical. If TRUE selects only one point per pixel}
}
\value{
data.frame containing longitude and latitude.
}
\description{
A function to remove the points that are outside the extension of the raster and to maintain at most one register per pixel.
}
\details{
Used internally in function \code{\link[modleR]{setup_sdmdata}}.
}
\examples{
occs <- coordenadas[[1]]
clean(occurrences = occs, predictors = example_vars,
      clean_dupl = TRUE,
      clean_nas = TRUE,
      clean_uni = TRUE)

}
\seealso{
\code{\link[raster]{cellFromXY}}, \code{\link[raster]{mask}}, \code{\link[raster]{extract}}
}
