% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean the occurrence records}
\usage{
clean(occurrences, predictors, clean_dupl = TRUE, clean_nas = TRUE,
  clean_uni = TRUE)
}
\arguments{
\item{occurrences}{data.frame with the coordinates}

\item{predictors}{an object with the rasters to be cut. Accepts an object of type RasterStack, generated by the function \code{\link[raster]{stack}}}
}
\value{
data.frame containing longitude and latitude.
}
\description{
A function to remove the points that are outside the extension of the raster and to maintain, at most one register per pixel.
}
\details{
Used internally in function \code{\link[modelos]{modelos}}.
}
\examples{
predictors <- example_vars
clean(occurrences = coordenas[,-1], predictors = predictors)

}
\seealso{
\code{\link[raster]{cellFromXY}}, \code{\link[raster]{mask}}, \code{\link[raster]{extract}}
}
\author{
Diogo S. B. Rocha
}
