% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean the occurrence records.}
\usage{
clean(occurrences, predictors, clean_dupl = FALSE, clean_nas = FALSE,
  clean_uni = FALSE)
}
\arguments{
\item{occurrences}{A data frame with occurrence data}

\item{predictors}{an object with the rasters to be cut. Accepts an object of type RasterStack, generated by the function \code{\link[raster]{stack}}}

\item{clean_dupl}{logical. If TRUE removes points with the same longitude and latitude}

\item{clean_nas}{logical. If TRUE removes points that are outside the bounds of the raster}

\item{clean_uni}{logical. If TRUE selects only one point per pixel}
}
\value{
data.frame containing longitude and latitude.
}
\description{
A function to remove the points that are outside the extension of the raster and to maintain, at most one register per pixel.
}
\details{
Used internally in function \code{\link[modleR]{setup_sdmdata}}.
}
\examples{
sp <- unique(coordenadas$sp)[1]
sp.coord <- coordenadas[coordenadas$sp == sp, -1]
clean(occurrences = sp.coord, predictors = example_vars)

}
\seealso{
\code{\link[raster]{cellFromXY}}, \code{\link[raster]{mask}}, \code{\link[raster]{extract}}
}
\author{
Diogo S. B. Rocha
}
