% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coordenadas}
\alias{coordenadas}
\title{Ocurrence points for four species in the Brazilian Atlantic Forest.}
\format{A list of 4: 
\describe{
  \item{Abarema_langsdorffii}{104 observations of 3 variables}
  \item{Eugenia_florida}{341 observations of  3 variables}
  \item{Leandra_carassana}{82 observations of 3 variables}
  \item{Ouratea_semiserrata}{90 observations of 3 variables}
}}
\usage{
coordenadas
}
\description{
A list with four elements containing ccurrence points from four species: Abarema langsdorfii, 
Eugenia florida, Leandra carassana and Ouratea semiserrata. 
Each element contains three variables: sp species name separated by _, lat contains Latitude information and lon contains
Longitude information.
}
\keyword{datasets}
