% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coordenadas}
\alias{coordenadas}
\title{Pontos de ocorrência de 4 espécies}
\format{Um data frame contendo 618 linhas e 3 colunas:
\describe{
  \item{sp}{Nome da espécie}
  \item{lat}{Latitude}
  \item{lon}{Longitude}
}}
\usage{
coordenadas
}
\description{
Uma tabela que contém 618 pontos de ocorrência de 4 espécies
}
\keyword{datasets}
