% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coordenadas}
\alias{coordenadas}
\title{Ocurrence points for four species in the Brazilian Atlantic Forest}
\format{A list of 4:
\describe{
  \item{Abarema_langsdorffii}{104 observations of 3 variables}
  \item{Eugenia_florida}{341 observations of  3 variables}
  \item{Leandra_carassana}{82 observations of 3 variables}
  \item{Ouratea_semiserrata}{90 observations of 3 variables}
}}
\usage{
coordenadas
}
\description{
A list with four elements containing occurrence points for four species:
\emph{Abarema langsdorfii}, \emph{Eugenia florida}, \emph{Leandra carassana}
and \emph{Ouratea semiserrata}.
Each element contains three variables: \code{sp} (species name separated by
"_"), \code{lat} and \code{lon} (Latitude and Longitude in decimal degrees)
}
\keyword{datasets}
