% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffer.R
\name{create_buffer}
\alias{create_buffer}
\title{Samples pseudoabsences inside a geographic buffer}
\usage{
create_buffer(occurrences, buffer_type = NULL, predictors,
  dist_buf = NULL, dist_min = NULL)
}
\arguments{
\item{occurrences}{A data frame with occurrence data. It should contain only two columns:
lon and lat, in that order.}

\item{buffer_type}{Character string indicating whether the buffer should be
calculated using the mean, median or maximum distance between occurrence points}

\item{predictors}{A RasterStack of predictor variables}

\item{dist_buf}{Optional, a distance in km for the inclusion buffer. If set it will override buffer_type}

\item{dist_min}{Optional, a distance in km for the exclusion buffer. ###If set it will override buffer_type}
}
\value{
Table of pseudoabsence points sampled within the selected distance

A buffer around the occurrence points
}
\description{
Samples pseudoabsences inside a geographic buffer
}
\details{
The sampling is performed by dismo::randomPoints() excluding the presence points (exclupep =TRUE)
}
\examples{
library(raster)
library(dplyr)
species <- sort(unique(coordenadas$sp))
occs <- coordenadas \%>\% filter(sp == species[1]) \%>\% dplyr::select(lon, lat)
buf <- create_buffer(occs, "mean", example_vars)
plot(buf)

}
\references{
VanDerWal, J., Shoo, L. P., Graham, C., & Williams, S. E. (2009). Selecting pseudo-absence data for presence-only distribution modeling: How far should you stray from what you know? Ecological Modelling, 220(4), 589-594. doi:10.1016/j.ecolmodel.2008.11.010
}
\seealso{
\code{\link[raster]{buffer}}

\code{\link[dismo]{randomPoints}}
}
\author{
Felipe Barros

Andrea Sánchez-Tapia

Fabrício Vilasboas
}
