% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffer.R
\name{create_buffer}
\alias{create_buffer}
\title{Samples pseudoabsences inside a geographic buffer.}
\usage{
create_buffer(species_name, occurrences, lon = "lon", lat = "lat",
  predictors, buffer_type = "median", dist_buf = NULL,
  env_buffer = FALSE, env_distance = "centroid", dist_min = NULL,
  max_env_dist = 0.5, buffer_shape, models_dir = "./models",
  write_buffer = F)
}
\arguments{
\item{species_name}{A character string with the species name. Because species name will be used as
a directory name, avoid non-ASCII characters, spaces and punctuation marks.
Recommendation is to adopt "Genus_species" format. See names in \code{\link{coordenadas}} as an example.}

\item{occurrences}{A data frame with occurrence data. Data must have at least columns with latitude an longitude values of species occurrences.
See \code{coordenadas} as an example.}

\item{lon}{The name of the longitude column. Defaults to "lon".}

\item{lat}{The name of the latitude column. Defaults to "lat".}

\item{predictors}{An object with the rasters to be cut.
Accepts an object of type RasterStack, generated by the function \code{\link[raster]{stack}}.}

\item{buffer_type}{Character string indicating whether the buffer should be
calculated using the "mean", "median", "maximum" distance between occurrence
points, or an absolute geographic "distance". If NULL pseudoabsences are randomly generated in the entire area
of the RasterStack of predictor variables.
filled with predictors. If set to "distance", "dist_buf" needs to
be specified. If set to "user", "buffer_shape" needs to be specified.}

\item{dist_buf}{Defines the width of the buffer. Needs to be specified if buffer_type = "distance".
Distance unit is in the same unit of the RasterStack of predictor variables}

\item{env_buffer}{Logical. Should an euclidean environmental filter be executed? If TRUE, \code{"env_distance"} and \code{"max_env_dist"} need to be specified.}

\item{env_distance}{Character. Type of environmental distance \code{"centroid", "mindist"}. The default is \code{"centroid"}.}

\item{dist_min}{Optional, a distance for the exclusion buffer
Distance unit is in the same unit of the RasterStack of predictor variables}

\item{max_env_dist}{Numeric. Quantile set as maximum environmental distance. Needs to be specified if env_buffer = TRUE}

\item{buffer_shape}{User-defined buffer shapefile in which pseudoabsences will be generated.
Needs to be specified if buffer_type = "user"}

\item{models_dir}{Folder path to save the output files.}

\item{write_buffer}{Logical. Should the resulting raster file be written? defaults to FALSE}
}
\value{
Table of pseudoabsence points sampled within the selected distance

A buffer around the occurrence points
}
\description{
Samples pseudoabsences inside a geographic buffer.
}
\details{
It will return a raster object with the same resolution and extent, and cropped by the predictors stack.
}
\examples{
library(raster)
sp <- names(coordenadas)[1]
occs <- coordenadas[[1]]
buf <- create_buffer(species_name=sp, occurrences=occs, predictors=example_vars)
plot(buf)

}
\references{
VanDerWal, J., Shoo, L. P., Graham, C., & Williams, S. E. (2009). Selecting pseudo-absence data for presence-only distribution modeling: How far should you stray from what you know? Ecological Modelling, 220(4), 589-594. doi:10.1016/j.ecolmodel.2008.11.010
}
\seealso{
\code{\link[raster]{buffer}}

\code{\link[dismo]{randomPoints}}
}
\author{
Felipe Barros

Andrea Sánchez-Tapia

Diogo S.B. Rocha

Sara Mortara
}
