% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffer.R
\name{create_buffer}
\alias{create_buffer}
\title{Samples pseudoabsences inside a geographic or environmental buffer}
\usage{
create_buffer(species_name, occurrences, lon = "lon", lat = "lat",
  predictors, buffer_type = "none", buffer_shape, dist_buf = NULL,
  env_filter = FALSE, env_distance = "centroid", min_env_dist = NULL,
  min_geog_dist = NULL, models_dir = "./models", write_buffer = FALSE)
}
\arguments{
\item{species_name}{A character string with the species name. Because species
name will be used as a directory name, avoid non-ASCII characters, spaces and
punctuation marks.
Recommendation is to adopt "Genus_species" format. See names in
\code{\link{example_occs}} as an example}

\item{occurrences}{A data frame with occurrence data. Data must have at least
 columns with latitude and longitude values of species occurrences.
See \code{\link{example_occs}} as an example}

\item{lon}{The name of the longitude column. Defaults to "lon"}

\item{lat}{The name of the latitude column. Defaults to "lat"}

\item{predictors}{A Raster or RasterStack object with the environmental
raster layers}

\item{buffer_type}{Character string indicating whether the buffer should be
calculated using the "\code{mean}", "\code{median}", "\code{maximum}"
distance between occurrence points, or an absolute geographic
"\code{distance}". If set to "\code{user}", a user-supplied shapefile will be
used as a sampling area, and \code{buffer_shape} needs to be specified. If
NULL, no distance buffer is applied. If set to "\code{distance}",
\code{dist_buf} needs to be specified}

\item{buffer_shape}{User-defined buffer shapefile in which pseudoabsences
will be generated. Needs to be specified if \code{buffer_type = "user"}}

\item{dist_buf}{Defines the width of the buffer. Needs to be specified if
\code{buffer_type = "distance"}. Distance unit is in the same unit of the
RasterStack of predictor variables}

\item{env_filter}{Logical. Should an euclidean environmental filter be
applied? If TRUE, \code{env_distance} and
\code{min_env_dist} need to be specified. Areas closest than
\code{min_env_dist} (expressed in quantiles in the environmental space)will
be omitted from the pseudoabsence sampling}

\item{env_distance}{Character. Type of environmental distance, either
"\code{centroid}" or "\code{mindist}". Defaults to "\code{centroid}", the
distance of each raster pixel to the environmental centroid of the
distribution. When set to "\code{mindist}", the minimum distance of each
raster pixel to any of the occurrence points is calculated. Needs to be
specified if \code{env_filter = TRUE}. A minimum value needs to be
specified (parameter \code{min_env_dist})}

\item{min_env_dist}{Numeric. Sets a minimum value to exclude the areas
closest (in the environmental space) to the occurrences or their centroid,
expressed in quantiles, from 0 (the closest) to 1. Defaults to 0.05,
excluding areas belonging to the 5% closest environmental values. Note that
since this is based on quantiles, and environmental similarity can take large
negative values, this is an abitrary value}

\item{min_geog_dist}{Optional, numeric. A distance for the exclusion of the
areas closest to the occurrence points (in the geographical space). Distance
 unit is in the same unit of the RasterStack of predictor variables}

\item{models_dir}{Folder path to save the output files. Defaults to
"\code{./models}"}

\item{write_buffer}{Logical. Should the resulting buffer RasterLayer be
written? Defaults to FALSE}
}
\value{
Table of pseudoabsence points sampled within the selected distance

A RasterLayer object containing the final buffer around the
occurrence points
}
\description{
This function is used internally by function \code{\link{setup_sdmdata}} to
define the area where pseudoabsences will be sampled in different ways.
First, it can create an inclusion buffer, within which the pseudoabsences
will be sampled, to restrict model evaluation to accesible areas. This can be
performed by either setting a user-defined shapefile, or by drawing a buffer
around the species occurrences: a geographic distance fixed value, or the
mean, median or maximum pairwise distance between occurrences. In addition to
this, an euclidean environmental distance filter can be superimposed to the
previous step, to control for overfitting by excluding areas that are too
close to the occurrence points, either in the environmental space or in the
geographic space.
The function will return the resulting buffer as a RasterStack object with
the same resolution and NA values of the predictors RasterStack.
}
\examples{
library(raster)
sp <- names(example_occs)[1]
occs <- example_occs[[1]]
buf <- create_buffer(species_name = sp,
                     occurrences = occs,
                     predictors = example_vars)
plot(buf)

}
\references{
VanDerWal, J., Shoo, L. P., Graham, C., & Williams, S. E. (2009).
Selecting pseudo-absence data for presence-only distribution modeling: How
far should you stray from what you know? Ecological Modelling, 220(4),
589-594. doi:10.1016/j.ecolmodel.2008.11.010
}
\seealso{
\code{\link[raster]{buffer}}  in \pkg{raster} package

\code{\link[dismo]{randomPoints}}  in \pkg{dismo} package
}
