% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_buffer.R
\name{create_buffer}
\alias{create_buffer}
\title{Samples pseudoabsences inside a geographic buffer}
\usage{
create_buffer(occurrences, buffer_type = NULL, predictors,
  dist_buf = NULL, buffer_shape = NULL, models_dir = "./models",
  species_name, write_buffer = F, ...)
}
\arguments{
\item{occurrences}{A data frame with occurrence data}

\item{buffer_type}{Character string indicating whether the buffer should be
calculated using the "mean", "median", "maximum" distance between occurrence points, or an absolute "distance". If set to "distance",
        "dist_buf" needs to be specified. If set to "user", "buffer_shape" needs to be specified.}

\item{predictors}{A RasterStack of predictor variables}

\item{dist_buf}{Defines the width of the buffer. Needs to be specified if buffer_type = "distance"}

\item{buffer_shape}{User-defined buffer shapefile. Needs to be specified if buffer_type = "user"}

\item{models_dir}{Folder path to save the output files}

\item{species_name}{A character string with the species name}

\item{write_buffer}{Logical. Should the resulting raster file be written? defaults to FALSE}

\item{...}{Other parameters from writeRaster}
}
\value{
Table of pseudoabsence points sampled within the selected distance

A buffer around the occurrence points
}
\description{
Samples pseudoabsences inside a geographic buffer
}
\details{
The sampling is performed by dismo::randomPoints() excluding the presence points (exclupep =TRUE)
}
\examples{
library(raster)
library(dplyr)
species <- sort(unique(coordenadas$sp))
occs <- coordenadas \%>\% filter(sp == species[1]) \%>\% dplyr::select(lon, lat)
buf <- create_buffer(occs, "mean", example_vars)
plot(buf)

}
\references{
VanDerWal, J., Shoo, L. P., Graham, C., & Williams, S. E. (2009). Selecting pseudo-absence data for presence-only distribution modeling: How far should you stray from what you know? Ecological Modelling, 220(4), 589-594. doi:10.1016/j.ecolmodel.2008.11.010
}
\seealso{
\code{\link[raster]{buffer}}

\code{\link[dismo]{randomPoints}}
}
\author{
Felipe Barros

Andrea Sánchez-Tapia

Fabrício Vilasboas
}
