% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_model.R
\name{crop_model}
\alias{crop_model}
\title{Crops and masks environmental raster based on specific mask}
\usage{
crop_model(modelo, mask)
}
\arguments{
\item{modelo}{RasterStack object of the ecological niche model}

\item{mask}{A SpatialPolygonsDataFrame to be used to mask the final models}
}
\value{
Croppped environmental raster based on specific buffer
}
\description{
Crops and masks environmental raster based on specific mask
}
\seealso{
\code{\link[raster]{crop}}

\code{\link[raster]{mask}}
}
