% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_model.R
\name{crop_model}
\alias{crop_model}
\title{Crops environmental raster based on specific buffer}
\usage{
crop_model(modelo, mask)
}
\arguments{
\item{modelo}{Raster with the ecological niche model.}

\item{mask}{A SpatialPolygonsDataFrame to be used to mask the final models.}
}
\value{
Croppped environmental raster based on specific buffer
}
\description{
Crops environmental raster based on specific buffer
}
\seealso{
\code{\link[raster]{crop}}

\code{\link[raster]{mask}}
}
