% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_any.R
\name{do_any}
\alias{do_any}
\title{Model fitting and predicting of ecological niche models using several algorithms}
\usage{
do_any(species_name, predictors, models_dir = "./models",
  algo = c("bioclim"), project_model = FALSE,
  proj_data_folder = "./data/proj", mask = NULL, write_png = FALSE,
  write_bin_cut = FALSE, threshold = "spec_sens", conf_mat = TRUE,
  equalize = TRUE)
}
\arguments{
\item{species_name}{A character string with the species name. Because species name will be used as 
a directory name, avoid non-ASCII characters, spaces and ponctuation marks. 
Recommendation is to adopt "Genus_species" format. See names in \code{\link{coordenadas}} as an example.}

\item{predictors}{An object with the rasters to be cut. 
Accepts an object of type RasterStack, generated by the function \code{\link[raster]{stack}}.}

\item{models_dir}{Folder path to save the output files.}

\item{algo}{The algorithm to be fitted \code{c("bioclim", "maxent", "domain",
"mahal", "glm", "svmk", "svme",
 "rf", "brt", "mindist", "centroid")}.}

\item{project_model}{Logical, whether to perform a projection.}

\item{proj_data_folder}{Path to projections -containing one or more
folders with the projection datasets, ex. "./env/proj/proj1".}

\item{mask}{A SpatialPolygonsDataFrame to be used to mask the final models.}

\item{write_png}{Logical, whether png files will be written.}

\item{write_bin_cut}{Logical, whether binary and cut model files(.tif, .png) should be written.}

\item{threshold}{Character string indicating threshold (cut-off) to transform model predictions 
to a binary score.
as in \code{\link[dismo]{threshold}}: "kappa", "spec_sens", "no_omission", "prevalence", 
"equal_sens_spec", "sensitivity". Default value is "spec_sens".}

\item{conf_mat}{Logical, whether confusion tables should be written in the HD.}

\item{equalize}{Logical, whether the number of presences and absences should be
equalized in randomForest and brt.}
}
\value{
A data frame with the evaluation statistics (TSS, AUC, etc).
}
\description{
This function reads the output from \code{\link{setup_sdmdata}} and 
computes ecological niche models for a species based on algorithm specified by the user. See details for
a description of all algorithms supported in this package.
}
\details{
See bellow for a description on the implementation of the algorithms supported in this package.
\describe{
\item{Bioclim}{Specified by \code{algo="bioclim"} uses \code{\link[dismo]{bioclim}} function in dismo
package \insertCite{hijmans_dismo:_2017}{modleR}. Bioclim is the  climate-envelope-model implemented by Henry Nix 
\insertCite{nix1986biogeographic}{modleR}, the first species 
distribuition modelling package. It is based on climate interpolation methods and despite its limiations 
it is still used in ecological niche modeling, specially for exploration and teaching purposes
\insertCite{@see also @booth_bioclim:_2014}{modleR}. 
In this package it is implemented by the function \code{\link[dismo]{bioclim}}, evaluated and predicted 
using \code{\link[dismo]{evaluate}} and \code{\link[dismo]{predict}} also from dismo package.  
}
\item{Maximum Entropy (Maxent)}{Specified either by \code{algo="maxent"} or \code{algo="maxnet"} 
corresponding to implementation by dismo \insertCite{hijmans_dismo:_2017}{modleR} and maxnet
\insertCite{maxnet}{modleR} packages respectivelly. Maxent is a machine learning method for modeling
species distributions based in incomplete data allowing ENM with presence-only data 
\insertCite{phillips_maximum_2006}{modleR}. If \code{algo="maxent"} model is fitted by the function
\code{\link[dismo]{maxent}}, evaluated and predicted using  \code{\link[dismo]{evaluate}} and 
\code{\link[dismo]{predict}} also in dismo package. If \code{algo="maxnet"} model is fitted by the 
function \code{\link[maxnet]{maxnet}} from maxnet package, evaluated using \code{\link[dismo]{evaluate}}
from dismo package with argument \code{type="logistic"} and predicted using \code{\link[raster]{predict}}
function from raster package. 
}
\item{Mahalanobis}{Specified by \code{algo="mahal"} uses \code{\link[dismo]{mahal}} function from dismo 
package. Corresponds to a distribution model based on Mahalanobis distance, a measure of the distance 
between a point P and a distribution D \insertCite{mahalanobis_generalized_1936}{modleR}. In this package 
it is implemented by the function \code{\link[dismo]{mahal}}, evaluated and predicted 
using \code{\link[dismo]{evaluate}} and \code{\link[dismo]{predict}} also from dismo package. 
}
\item{Domain}{Specified by \code{algo="domain"} uses \code{\link[dismo]{domain}} function from dismo
package. Computes  point-to-point similarity based on Gower distance between environmental variables 
\insertCite{carpenter_domain:_1993}{modleR}. \insertCite{hijmans_dismo:_2017}{modleR} state that 
one should use it with caution because it does not perform well compared to other algorithms
\insertCite{elith_novel_2006,hijmans_ability_2006}{modleR}. 
We add that it is a slow algorithm.
In this package it is implemented by the function \code{\link[dismo]{domain}}, evaluated and predicted 
using \code{\link[dismo]{evaluate}} and \code{\link[dismo]{predict}} also from dismo package.
}
\item{Support Vector Machines (SVM)}{
}
\item{GLM}{
}
\item{Random Forest}{
}
\item{Euclidian algorithms}{
to do or not to do
}
\item{Boosted Refression Trees (BRT)}{
}
}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[dismo]{bioclim}}

\code{\link[dismo]{maxent}}

\code{\link[maxnet]{maxnet}}

\code{\link[dismo]{domain}}

\code{\link[dismo]{mahal}}

\code{\link[dismo]{evaluate}}

\code{\link[dismo]{predict}}

\code{\link[raster]{predict}}
}
\author{
Andrea Sánchez-Tapia
}
