% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_any.R
\name{do_any}
\alias{do_any}
\title{Fits ecological niche models using several algorithms}
\usage{
do_any(species_name, predictors, models_dir = "./models",
  algo = c("bioclim"), project_model = FALSE,
  proj_data_folder = "./data/proj", mask = NULL, write_png = FALSE,
  write_bin_cut = FALSE, threshold = "spec_sens", conf_mat = TRUE,
  equalize = TRUE)
}
\arguments{
\item{species_name}{A character string with the species name. Because species name will be used as 
a directory name, avoid non-ASCII characters, spaces and ponctuation marks. 
Recommendation is to adopt "Genus_species" format. See names in \code{\link{coordenadas}} as an example.}

\item{predictors}{An object with the rasters to be cut. 
Accepts an object of type RasterStack, generated by the function \code{\link[raster]{stack}}.}

\item{models_dir}{Folder path to save the output files.}

\item{algo}{The algorithm to be fitted \code{c("bioclim", "maxent", "domain",
"mahal", "glm", "svmk", "svme",
 "rf", "brt", "mindist", "centroid")}.}

\item{project_model}{Logical, whether to perform a projection.}

\item{proj_data_folder}{Path to projections -containing one or more
folders with the projection datasets, ex. "./env/proj/proj1".}

\item{mask}{A SpatialPolygonsDataFrame to be used to mask the final models.}

\item{write_png}{Logical, whether png files will be written.}

\item{write_bin_cut}{Logical, whether binary and cut model files(.tif, .png) should be written.}

\item{threshold}{Character string indicating threshold (cut-off) to transform model predictions 
to a binary score.
as in \code{\link[dismo]{threshold}}: "kappa", "spec_sens", "no_omission", "prevalence", 
"equal_sens_spec", "sensitivity". Default value is "spec_sens".}

\item{conf_mat}{Logical, whether confusion tables should be written in the HD.}

\item{equalize}{Logical, whether the number of presences and absences should be
equalized in randomForest and brt.}
}
\value{
A data frame with the evaluation statistics (TSS, AUC, etc).
}
\description{
Fits ecological niche models using several algorithms
}
\details{
Biolclim algorithm (\code{algo="bioclim"}) uses \code{\link[dismo]{bioclim}} function in  dismo
package. Bioclim is climate-envelope-model implemented by Henry Nix 
\insertCite{nix1986biogeographic}{modleR}.
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[dismo]{bioclim}}

\code{\link[dismo]{maxent}}

\code{\link[dismo]{domain}}

\code{\link[dismo]{mahal}}
}
\author{
Andrea Sánchez-Tapia
}
