% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_model.R
\name{do_any}
\alias{do_any}
\title{Fits ecological niche models using any dismo algorithm.}
\usage{
do_any(species_name, coordinates, predictors, models_dir = "./models",
  algo = c("bioclim"), project_model = FALSE, projections = NULL,
  mask = NULL, write_png = FALSE, ...)
}
\arguments{
\item{species_name}{A character string with the species name}

\item{coordinates}{A data frame with occurrence data}

\item{predictors}{A RasterStack of predictor variables}

\item{models_dir}{Folder path to save the output files}

\item{algo}{The algorithm to be fitted \code{c("bioclim", "maxent","domain",
"mahal", "glm", "svm.k","svm.e","rf")}}

\item{project_model}{Logical, whether to perform a projection}

\item{projections}{The RasterStack of projeciton variables}

\item{mask}{A SpatialPolygonsDataFrame to be used to mask the final models}

\item{write_png}{Logical, whether png files will be written}

\item{...}{Any parameter from \link{setup_sdmdata}}
}
\value{
A data frame with the evaluation statistics (TSS, AUC, etc.)
}
\description{
Fits ecological niche models using any dismo algorithm.
}
\seealso{
\code{\link[dismo]{bioclim}}

\code{\link[dismo]{maxent}}

\code{\link[dismo]{domain}}

\code{\link[dismo]{mahal}}
}
\author{
Andrea Sánchez-Tapia
}
