% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_model.R
\name{do_any}
\alias{do_any}
\title{Fits ecological niche models using any dismo algorithm.}
\usage{
do_any(species_name, occurrences, predictors, models_dir = "./models",
  algo = c("bioclim"), project_model = FALSE,
  proj_data_folder = "./data/proj", mask = NULL, write_png = FALSE,
  write_bin_cut = TRUE, buffer_type = NULL, dist_buf = NULL,
  conf_mat = TRUE, equalize = TRUE, ...)
}
\arguments{
\item{species_name}{A character string with the species name}

\item{occurrences}{A data frame with occurrence data}

\item{predictors}{A RasterStack of predictor variables}

\item{models_dir}{Folder path to save the output files}

\item{algo}{The algorithm to be fitted \code{c("bioclim", "maxent","domain",
"mahal", "glm", "svm.k","svm.e","rf")}}

\item{project_model}{Logical, whether to perform a projection}

\item{proj_data_folder}{the path to projections -containing one or more
folders with the projection datasets, ex. "./env/proj/proj1"}

\item{mask}{A SpatialPolygonsDataFrame to be used to mask the final models}

\item{write_png}{Logical, whether png files will be written}

\item{write_bin_cut}{Logical, whether binary and cut model files(.tif, .png) should be written}

\item{buffer_type}{Character string indicating whether the buffer should be
calculated using the "mean", "median", "maximum" distance between occurrence points, or an absolute "distance". If set to "distance",
        "dist_buf" needs to be specified. If set to "user", "buffer_shape" needs to be specified.}

\item{dist_buf}{Defines the width of the buffer. Needs to be specified if buffer_type = "distance"}

\item{conf_mat}{Logical, whether confusion tables should be written in the HD}

\item{equalize}{Logical, whether the number of presences and absences should be
equalized in randomForest and brt.}

\item{...}{Any parameter from \link{setup_sdmdata}}
}
\value{
A data frame with the evaluation statistics (TSS, AUC, etc.)
}
\description{
Fits ecological niche models using any dismo algorithm.
}
\seealso{
\code{\link[dismo]{bioclim}}

\code{\link[dismo]{maxent}}

\code{\link[dismo]{domain}}

\code{\link[dismo]{mahal}}
}
\author{
Andrea Sánchez-Tapia
}
