% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/any_model.R
\name{do_any}
\alias{do_any}
\title{Fits ecological niche models using several algorithms.}
\usage{
do_any(species_name, sdmdata, predictors, models_dir = "./models",
  algo = c("bioclim"), project_model = FALSE,
  proj_data_folder = "./data/proj", mask = NULL, write_png = FALSE,
  write_bin_cut = FALSE, threshold = "spec_sens", conf_mat = TRUE,
  equalize = TRUE, ...)
}
\arguments{
\item{species_name}{A character string with the species name}

\item{sdmdata}{sdmdata object resulting from \link{setup_sdmdata}}

\item{algo}{The algorithm to be fitted \code{c("bioclim", "maxent", "domain",
"mahal", "glm", "svmk", "svme",
 "rf", "brt", "mindist", "centroid")}}

\item{project_model}{Logical, whether to perform a projection}

\item{proj_data_folder}{the path to projections -containing one or more
folders with the projection datasets, ex. "./env/proj/proj1"}

\item{mask}{A SpatialPolygonsDataFrame to be used to mask the final models}

\item{write_png}{Logical, whether png files will be written}

\item{write_bin_cut}{Logical, whether binary and cut model files(.tif, .png) should be written}

\item{conf_mat}{Logical, whether confusion tables should be written in the HD}

\item{equalize}{Logical, whether the number of presences and absences should be
equalized in randomForest and brt.}

\item{...}{Any parameter from \link{setup_sdmdata}}
}
\value{
A data frame with the evaluation statistics (TSS, AUC, etc.)
}
\description{
Fits ecological niche models using several algorithms.
}
\seealso{
\code{\link[dismo]{bioclim}}

\code{\link[dismo]{maxent}}

\code{\link[dismo]{domain}}

\code{\link[dismo]{mahal}}
}
\author{
Andrea Sánchez-Tapia
}
