% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_enm.R
\name{do_enm}
\alias{do_enm}
\title{Fits ecological niche models for various algorithms.}
\usage{
do_enm(sp, coordinates, partitions, buffer = FALSE, seed = 512, predictors,
  models.dir = "./models", project.model = FALSE, projections = NULL,
  mask = NULL, write_png = FALSE, n.back, bioclim = FALSE,
  domain = FALSE, glm = FALSE, mahal = FALSE, maxent = FALSE,
  rf = FALSE, svm = FALSE, svm2 = FALSE)
}
\arguments{
\item{sp}{A character string with the species name}

\item{coordinates}{A two-column data frame with the occurrence points}

\item{partitions}{The number of partitions for a cross validation}

\item{buffer}{Defines if a buffer will be used to sample pseudo-absences
(F, "mean", "median", "max")}

\item{seed}{For reproducibility purposes}

\item{predictors}{A RasterStack of predictor variables}

\item{models.dir}{Folder path to save the output files}

\item{project.model}{Logical, whether to perform a projection}

\item{projections}{The RasterStack of projeciton variables}

\item{mask}{A SpatialPolygonsDataFrame to be used to mask the final models}

\item{write_png}{Logical, whether png files will be written}

\item{n.back}{Number of pseudoabsence points}

\item{bioclim}{Execute bioclim from the dismo implementation}

\item{domain}{Execute domain from the dismo implementation}

\item{glm}{Execute GLM as suggested by the dismo documentation}

\item{mahal}{Execute mahalanobis distance from the dismo implementation}

\item{maxent}{Execute maxent from the dismo implementation}

\item{rf}{Execute random forests from randomForest() as suggested by the dismo documentation}

\item{svm}{Execute svm from kernlab package}

\item{svm2}{Execute svm from e1071 package}
}
\value{
A set of ecological niche models for each partition and algorithm,
        written in the \code{models.dir} subfolder
}
\description{
Fits ecological niche models for various algorithms.
}
