% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm2.R
\name{do_SVM2}
\alias{do_SVM2}
\title{Fits ecological niche models using SVM from package e1071}
\usage{
do_SVM2(sp, coordinates, partitions, buffer, seed, predictors, models.dir,
  project.model, projections, mask, write_png, n.back)
}
\arguments{
\item{sp}{A character string with the species name}

\item{coordinates}{A two-column data frame with the occurrence points}

\item{buffer}{Defines if a buffer will be used to sample pseudo-absences (F, "mean", "median", "max")}

\item{seed}{For reproducibility purposes}

\item{predictors}{A RasterStack of predictor variables}

\item{models.dir}{Folder path to save the output files}

\item{mask}{A SpatialPolygonsDataFrame to be used as a mask to cut the final models}

\item{write_png}{Logical, whether png files will be written}

\item{n.back}{Number of pseudoabsence points}
}
\value{
A data frame with the evaluation statistics (TSS, AUC, and their respective thresholds)
}
\description{
Fits ecological niche models using SVM from package e1071
}
