% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble}
\alias{ensemble}
\title{Gera um modelo só por espécie}
\usage{
ensemble(sp, models.dir = "./models", final.dir = "presfinal",
  ensemble.dir = "ensemble", occs = spp.filt,
  which.models = c("Final.bin.mean3", "Final.mean.bin7"), consensus = F,
  consensus.level = 0.5)
}
\arguments{
\item{sp}{Um nome de espécie}

\item{models.dir}{Path do diretório onde estão os modelos}

\item{final.dir}{Path para subdiretório escrito com finalModel()}

\item{ensemble.dir}{Path para os arquivos de saída}

\item{occs}{Pontos de ocorrência da espécie}

\item{which.models}{Qual tipo de modelo final será juntado (ex. bin.mean3)}

\item{consensus}{Se será usada uma regra de consenso para cortar o ensemble}

\item{consensus.level}{Quanto dos modelos será retido (0,5 = maioria)}
}
\description{
Gera um modelo só por espécie
}
