% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ensemble_model}
\alias{ensemble_model}
\title{Create ensemble models from several algorithms.}
\usage{
ensemble_model(species.name, occs, models.dir = "./models",
  final.dir = "final_models", ensemble.dir = "ensemble",
  which.models = c("final_model_3", "final_model_7"), consensus = FALSE,
  consensus.level = 0.5, write_png = T, write_raw_map = F)
}
\arguments{
\item{species.name}{A character string with the species name}

\item{occs}{A two-column data frame with the occurrence points}

\item{models.dir}{Character string. Folder path where the input files
are located}

\item{final.dir}{Character string, name of the subfolder where the files
for the final models are located}

\item{ensemble.dir}{Character string, name of the folder to save the output
files. A subfolder will be created.}

\item{which.models}{Which final_model() will be used? Currently it can be:
\describe{
  \item{\code{weighted_AUC} or \code{weighted_TSS}}{the models weighted
  by TSS or AUC}
  \item{\code{final_model_3}}{the binary model created by selecting or not the
   partitions, taking their mean and cutting by the mean threshold that
   maximizes TSS (or other dismo thresholds)}
  \item{\code{final_model_7}}{the mean of the selected binary models}
  \item{\code{final_model_8}}{the binary consensus from \code{final_model_7}}
}}

\item{consensus}{Logical. Will a consensus rule be applied?}

\item{consensus.level}{Threshold for the consensus rule, betwen 0 and 1
(0.5 means a majority rule).}

\item{write_png}{Write png? Defaults to TRUE}

\item{write_raw_map}{Create a mean raw map without margins}
}
\value{
A rasterStack with the minimun, maximum, median, mean and standard
deviation of the assembled models. A set of ensemble models and figures
(optional) written in the \code{ensemble.dir} subfolder
}
\description{
\code{ensemble} takes the input of \code{final_model} and builds an ensemble
 model by calculating the mean of each final model per algorithm.
}
\seealso{
\link{final_model}
}
