% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Mean distance to the centroid of a distribution
This function calculates the mean distance to the centroid of a distribution based either on mean values or using a minimmum distance algorithm}
\usage{
euclidean(predictors, occurrences, algo = "centroid", filename = "",
  ...)
}
\arguments{
\item{predictors}{An object with the rasters to be cut.
Accepts an object of type RasterStack, generated by the function \code{\link[raster]{stack}}.}

\item{occurrences}{A data frame with occurrence data. Data must have at least columns with latitude an longitude values of species occurrences.
See \code{coordenadas} as an example.}

\item{algo}{Character. Any in centroid" or "mindist"}

\item{filename}{Optional. The raster that will be created in disk}

\item{...}{other parameters in raster::writeRaster()}
}
\description{
Mean distance to the centroid of a distribution
This function calculates the mean distance to the centroid of a distribution based either on mean values or using a minimmum distance algorithm
}
\examples{
centr <- euclidean(example_vars, occurrences = coordenadas[[1]][,c(2,3)])
raster::plot(centr)
}
