% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Calculates environmental euclidean distances to the occurrences of a species}
\usage{
euclidean(predictors, occurrences, algo = "centroid", filename = "",
  ...)
}
\arguments{
\item{predictors}{An object with the rasters to be cut.
Accepts an object of type RasterStack, generated by the function \code{\link[raster]{stack}}.}

\item{occurrences}{A data frame with occurrence data. Data must have at least columns with latitude an longitude values of species occurrences.
See \code{coordenadas} as an example.}

\item{algo}{Character. Either "centroid" or "mindist"}

\item{filename}{Optional. The raster that will be created in disk}

\item{...}{other parameters in raster::writeRaster()}
}
\description{
This function calculates the euclidean distance from each pixel in a
RasterStack of predictor variables to the environmental centroid of a
distribution or the minimum distance from each pixel to any occurrence point
}
\examples{
centr <- euclidean(example_vars, occurrences = coordenadas[[1]][, c(2,3)])
raster::plot(centr)
}
