% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Calculates environmental euclidean distances to the occurrences of a species}
\usage{
euclidean(predictors, occurrences, env_dist = "centroid",
  filename = "", ...)
}
\arguments{
\item{predictors}{A Raster or RasterStack object with the environmental
raster layers}

\item{occurrences}{A data frame with occurrence data. Data must have at least
 columns with latitude and longitude values of species occurrences.
See \code{\link{example_occs}} as an example}

\item{env_dist}{Character. Either "\code{centroid}" or "\code{mindist}"}

\item{filename}{Optional. The raster that will be created in disk}

\item{...}{other parameters in \code{\link[raster]{writeRaster}}}
}
\description{
This function calculates the euclidean distance from each pixel in a
RasterStack of predictor variables to the environmental centroid of a
distribution (\code{centroid}) or the minimum distance from each pixel to
any of the occurrence points (\code{mindist})
}
\examples{
sp_coord <- coordenadas[[1]][, c("lon", "lat")]
centr <- euclidean(example_vars,
                   occurrences = sp_coord)
raster::plot(centr)
}
