% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_model.R
\name{final_model}
\alias{final_model}
\title{Joins ENM from several partitions, creating a model per algorithm.}
\usage{
final_model(sp, select.partitions = TRUE, algoritmos = c("bioclim",
  "domain", "glm", "mahal", "maxent", "rf", "svm"),
  threshold = c("spec_sens"), TSS.value = 0.7, models.dir = "./models",
  final.dir = "final_models")
}
\arguments{
\item{sp}{A character string with the species name}

\item{select.partitions}{TRUE ou FALSE}

\item{algoritmos}{Which algorithms will be processed.}

\item{threshold}{Which cutting threshold will be used, defaults to "spec_sens"}

\item{TSS.value}{Threshold to select models from TSS values}

\item{models.dir}{Character. Folder path where the input files are located}

\item{final.dir}{Character. Name of the folder to save the output files.
A subfolder will be created.}
}
\value{
A set of ecological niche models and figures (optional) written in the
         \code{final.dir} subfolder
}
\description{
Joins ENM from several partitions, creating a model per algorithm.
}
