% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_model.R
\name{final_model}
\alias{final_model}
\title{Joins ENM from several partitions, creating a model per algorithm.}
\usage{
final_model(species.name, select.partitions = TRUE,
  weight.partitions = FALSE, algorithms = NULL,
  threshold = c("spec_sens"), weight.par = c("TSS", "AUC"),
  TSS.value = 0.7, consensus.level = 0.5, models.dir = "./models",
  final.dir = "final_models")
}
\arguments{
\item{species.name}{A character string with the species name}

\item{select.partitions}{TRUE ou FALSE}

\item{algorithms}{Which algorithms will be processed.}

\item{threshold}{Which selecting threshold will be used to cut the mean
models in final_model_3 approach (see vignettes), it
defaults to "spec_sens" but any dismo threshold
can be used: "kappa", "no_omission", "prevalence",
"equal_sens_spec", "sensitivity".}

\item{TSS.value}{Threshold to select models from TSS values}

\item{models.dir}{Character. Folder path where the input files are located}

\item{final.dir}{Character. Name of the folder to save the output files.
A subfolder will be created.}
}
\value{
A set of ecological niche models and figures (optional) written in the
         \code{final.dir} subfolder
}
\description{
This function reads the output from dismo.mod and creates a model per species
}
