% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_model.R
\name{final_model}
\alias{final_model}
\title{Joins ENM from several partitions, creating a model per algorithm}
\usage{
final_model(species_name, algorithms = NULL, scale_models = TRUE,
  consensus_level = 0.5, models_dir = "./models",
  final_dir = "final_models", proj_dir = "present",
  which_models = c("raw_mean"), mean_th_par = c("spec_sens"),
  uncertainty = FALSE, png_final = TRUE, ...)
}
\arguments{
\item{species_name}{A character string with the species name. Because species
name will be used as a directory name, avoid non-ASCII characters, spaces and
punctuation marks.
Recommendation is to adopt "Genus_species" format. See names in
\code{\link{example_occs}} as an example}

\item{algorithms}{Character vector specifying which algorithms will be
processed. Note that it can have length > 1, ex. \code{c("bioclim", "rf")}.
Defaults to NULL: if no name is given it will process all algorithms present
in the evaluation files}

\item{scale_models}{Logical. Whether input models should be scaled between 0
and 1}

\item{consensus_level}{Which proportion of binary models will be kept when
creating \code{bin_consensus}}

\item{models_dir}{Character. Folder path where the input files are located}

\item{final_dir}{Character. Name of the folder to save the output files. A
subfolder will be created, defaults to "final_model"}

\item{proj_dir}{Character. The name of the subfolder with the projection.
Defaults to "present" but can be set according to the other projections (i.e.
to execute the function in projected models)}

\item{which_models}{Which \code{final_model} will be used? Currently it can
be:
\describe{
  \item{\code{raw_mean}}{Continuous model generated by the mean of the raw
  models (scale from 0 to 1)}
  \item{\code{raw_mean_th}}{Cuts the \code{raw_mean} by the mean of the
  threshold selected in \code{mean_th_par} to make a binary model}
   \item{\code{raw_mean_cut}}{Recovers \code{raw_mean} values above the
   threshold selected in \code{mean_th_par}. Generates a continuous model
   above this threshold while keeping null values below it}
  \item{\code{bin_mean}}{The mean of the binary models, created by cutting
  the raw mean models by the threshold selected in \code{mean_th_par}.
  Generates a model in a discrete scale (0 to 1 in 1/n intervals where n is
  the number of partitions)}
  \item{\code{bin_consensus}}{The binary consensus from \code{bin_mean}.
  Parameter \code{consensus_level} must be defined, 0.5 means a majority
  consensus}
}}

\item{mean_th_par}{A threshold that will be used to cut the raw mean models
if \code{which_models} is set to "\code{raw_mean_th}" or to transform the raw
mean models into binary models if \code{which_models} is set to "\code{bin_consensus}"
Defaults to "\code{spec_sens}" but any \pkg{dismo} threshold can be used (see
 function \code{\link[dismo]{threshold}}): "\code{kappa}",
 "\code{no_omission}", "\code{prevalence}", "\code{equal_spec_sens}",
 "\code{sensitivity}"
 @param sensitivity The sensitivity value for threshols "\code{sensitivity}".
 Defaults to 0.9}

\item{uncertainty}{Whether an uncertainty map, measured as range (max-min)
should be calculated}

\item{png_final}{Logical. If \code{TRUE}, writes png files of the final
models}

\item{...}{Other parameters from \code{\link[raster]{writeRaster}},
especially \code{overwrite = TRUE}, when needed}
}
\value{
Returns a data frame with final statistics of the partitions included
 in the final model

Writes on disk a data frame with mean statistics of the partitions
included in the final model

Writes on disk a set of ecological niche models (.tif files) in the
\code{final_dir} subfolder

If \code{png_final = TRUE} writes .png figures in the
\code{final_dir} subfolder
}
\description{
This function reads the models generated either by \code{\link{do_any}} or
\code{\link{do_many}} (i.e. one model per partition per algorithm) and
summarizes them into a final model for each species-algorithm combination.
These final models may be created from the raw continuous models or the
binary models (see \code{\link{do_any}}). From these means, several
outputs can be created, see \code{which_models} for details about
the final outputs available. The uncertainty between partitions, taken as
ranges (maximum - minimum values) between partitions may also be calculated.
Just as in \emph{no silver bullets in correlative ecological niche modeling},
no method for evaluating and selecting partitions is always better and these
are only a subset of the possibilities. The user should choose how to create
the final model based on their assumptions. We simply recommend to focus on
statistical clarity rather than significance
\insertCite{dushoff_can_2019}{modleR}.
}
\examples{
# run setup_sdmdata
sp <- names(example_occs)[1]
sp_coord <- example_occs[[1]]
sp_setup <- setup_sdmdata(species_name = sp,
                          occurrences = sp_coord,
                          predictors = example_vars)

# run do_any
sp_bioclim <- do_any(species_name = sp,
                     predictors = example_vars,
                     algorithm = "bioclim")

# run final_model
sp_final <- final_model(species_name = sp,
                        algorithms = "bioclim",
                        which_models = c("bin_consensus"),
                        consensus_level = 0.5,
                        overwrite = TRUE)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[dismo]{threshold}}  in \pkg{dismo} package

\code{\link[raster]{writeRaster}}  in \pkg{raster} package
}
