% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_model.R
\name{final_model}
\alias{final_model}
\title{Joins ENM from several partitions, creating a model per algorithm}
\usage{
final_model(species_name, algorithms = NULL, select_partitions = TRUE,
  select_par = "TSS", select_par_val = 0.7, weight_par = NULL,
  cut_level = c("spec_sens"), scale_models = TRUE,
  consensus_level = 0.5, models_dir = "./models",
  final_dir = "final_models", proj_dir = "present",
  which_models = c("raw_mean"), uncertainty = FALSE,
  write_final = TRUE, ...)
}
\arguments{
\item{species_name}{A character string with the species name. Because species
name will be used as a directory name, avoid non-ASCII characters, spaces and
punctuation marks.
Recommendation is to adopt "Genus_species" format. See names in
\code{\link{example_occs}} as an example}

\item{algorithms}{Character vector specifying which algorithms will be
processed. Note that it can have length > 1, ex. \code{c("bioclim", "rf")}. Defaults to
 NULL: if no name is given it will process all algorithms present in the evaluation files}

\item{select_partitions}{Logical. If TRUE only partitions above a particular
threshold value are selected}

\item{select_par}{Which performance statistic should be used to select the
partitions. Defaults to NULL but either \code{"AUC"} or \code{"TSS"} can be
used}

\item{select_par_val}{Performance metric value to select partitions}

\item{weight_par}{Which performance statistic should be used to weigh the
partitions. Defaults to NULL but either \code{c("AUC", "TSS")} can be used}

\item{cut_level}{Which selecting threshold will be used to cut the mean
models. Default is set to "\code{spec_sens}" but any \pkg{dismo} threshold (see
function \code{\link[dismo]{threshold}}) can be used: "\code{kappa}", "\code{no_omission}",
 "\code{prevalence}", "\code{equal_sens_spec}", "\code{sensitivity}"}

\item{scale_models}{Logical. Whether input models should be scaled between 0
and 1}

\item{consensus_level}{Which proportion of binary models will be kept when creating \code{bin_consensus}}

\item{models_dir}{Character. Folder path where the input files are located}

\item{final_dir}{Character. Name of the folder to save the output files.
A subfolder will be created, defaults to "final_model"}

\item{proj_dir}{Character. The name of the subfolder with the projection.
Defaults to "present" but can be set according to the other projections (i.e.
to execute the function in projected models)}

\item{which_models}{Which \code{final_model} will be used? Currently it can
be:
\describe{
  \item{\code{raw_mean}}{Continuous model generated by the mean of the raw
  models (scale from 0 to 1)}
  \item{\code{raw_mean_th}}{Cuts the \code{raw_mean} by the mean of the
  thresholds that maximize the selected evaluation metric (e.g. TSS
  (\code{spec_sens}) to make a binary model}
   \item{\code{raw_mean_cut}}{Recovers \code{raw_mean} values above the mean
   threshold that maximizes the selected evaluation metric (e.g. TSS
   (\code{spec_sens}) or other \pkg{dismo} thresholds). Generates a
   continuous model}
  \item{\code{bin_mean}}{The mean of the selected binary models. Generates a
  model in a discrete scale (0 to 1 in 1/n intervals where n is the number of retained models)}
  \item{\code{bin_consensus}}{The binary consensus from \code{bin_mean}.
  Parameter \code{consensus_level} must be defined, 0.5 means a majority
  consensus}
  \item{\code{cut_mean}}{The mean of the selected cut models. Values below
  the thresholds are down-weighted by zeros}
}}

\item{uncertainty}{Whether an uncertainty map, measured as range (max-min)
should be calculated}

\item{write_final}{Logical. If \code{TRUE}, writes png files of the final
models}

\item{...}{Other parameters from \code{\link[raster]{writeRaster}}, especially \code{overwrite = TRUE}, when needed}
}
\value{
Returns a data frame with final statistics of the partitions included in the final model

Writes on disk a set of ecological niche models (.tif files) in the \code{final_dir} subfolder

If \code{write_final = TRUE} writes .png figures
 in the \code{final_dir} subfolder
}
\description{
This function reads the models generated either by \code{\link{do_any}} or
\code{\link{do_many}} (i.e. one model per partition per algorithm) and
summarizes them into a final model for each species-algorithm combination.
These final models may be created from the raw uncut models, the binary
models or the "cut" models (see \code{\link{do_any}}). They may use all
partitions or select only partitions (\code{select_partitions = TRUE}) with performance metrics above a defined
value (specified in \code{select_par_val}). The function can also calculate weighted means of the partitions (selected or not) using
performance statistic ("\code{AUC}" or "\code{TSS}" given to \code{weight_par} argument). A majority rule can be applied to binary
models and uncertainty taken as ranges between partitions may also be
calculated. Analogous to \emph{no silver bullets in correlative ecological niche modeling}, no method for evaluating and selecting partitions is always better. The user should choose how to create the final model based on their assumptions and use. We simply recommend to focus on statistical clarity rather than significance \insertCite{dushoff_can_2019}{modleR}.
}
\examples{
# run setup_sdmdata
sp <- names(example_occs)[1]
sp_coord <- example_occs[[1]]
sp_setup <- setup_sdmdata(species_name = sp,
                          occurrences = sp_coord,
                          example_vars)

# run do_any
sp_bioclim <- do_any(species_name = sp,
                     predictors = example_vars,
                     algorithm = "bioclim")

# run final_model
sp_final <- final_model(species_name = sp,
                        algorithms = "bioclim",
                        select_partitions = TRUE,
                        select_par = "TSS",
                        select_par_val = 0,
                        which_models = c("bin_consensus"),
                        consensus_level = 0.5,
                        overwrite = TRUE)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[dismo]{threshold}}  in \pkg{dismo} package

\code{\link[raster]{writeRaster}}  in \pkg{raster} package
}
