% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_model.R
\name{final_model}
\alias{final_model}
\title{Joins ENM from several partitions, creating a model per algorithm.}
\usage{
final_model(species_name, algorithms = NULL, weight.par = NULL,
  select.partitions = TRUE, threshold = c("spec_sens"),
  select.par = "TSS", select.par.val = 0.7, consensus.level = 0.5,
  models.dir = "./models", final.dir = "final_models",
  which.models = c("final_model_3", "final_model_7"), write_png = T)
}
\arguments{
\item{species_name}{A character string with the species name}

\item{algorithms}{Which algorithms will be processed. If no name is given it
will process all algorithms present in the evaluation files}

\item{weight.par}{Which performance statistic should be used to weight the
partitions. Defaults to NULL but either \code{c("AUC", "TSS")} can be used.}

\item{select.partitions}{TRUE ou FALSE}

\item{threshold}{Which selecting threshold will be used to cut the mean
models in final_model_3 approach (see vignettes), it
defaults to "spec_sens" but any dismo threshold
can be used: "kappa", "no_omission", "prevalence",
"equal_sens_spec", "sensitivity".}

\item{select.par}{Which performance statistic should be used to select the
partitions- Defaults to NULL but either \code{c("AUC", "TSS")} can be used.}

\item{select.par.val}{Threshold to select models from TSS values}

\item{consensus.level}{Which proportion of models will be kept when creating
\code{final_model_8} (binary)}

\item{models.dir}{Character. Folder path where the input files are located}

\item{final.dir}{Character. Name of the folder to save the output files.
                 A subfolder will be created.
\describe{
  \item{\code{weighted_AUC} or \code{weighted_TSS}}{the models weighted
  by TSS or AUC}
  \item{\code{final_model_3}}{the binary model created by selecting or not the
   partitions, taking their mean and cutting by the mean threshold that
   maximizes TSS (or other dismo thresholds)}
  \item{\code{final_model_7}}{the mean of the selected binary models}
  \item{\code{final_model_8}}{the binary consensus from \code{final_model_7}}
}}

\item{which.models}{Which final_model() will be used? Currently it can be:}

\item{write_png}{Writes png files of the final models}
}
\value{
A set of ecological niche models and figures (optional) written in the
         \code{final.dir} subfolder
}
\description{
This function reads the output from dismo.mod and creates a model per species
}
