% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_model.R
\name{final_model}
\alias{final_model}
\title{Joins ENM from several partitions, creating a model per algorithm}
\usage{
final_model(species_name, algorithms = NULL, weight_par = NULL,
  select_partitions = TRUE, cut_level = c("spec_sens"),
  scale_models = TRUE, select_par = "TSS", select_par_val = 0.7,
  consensus_level = 0.5, models_dir = "./models",
  final_dir = "final_models", proj_dir = "present",
  which_models = c("raw_mean"), uncertainty = FALSE,
  write_final = TRUE, ...)
}
\arguments{
\item{species_name}{A character string with the species name. Because species name will be used as
a directory name, avoid non-ASCII characters, spaces and punctuation marks.
Recommendation is to adopt "Genus_species" format. See names in \code{\link{coordenadas}} as an example.}

\item{algorithms}{Which algorithms will be processed. If no name is given it
will process all algorithms present in the evaluation files.}

\item{weight_par}{Which performance statistic should be used to weigh the
partitions. Defaults to NULL but either \code{c("AUC", "TSS")} can be used.}

\item{select_partitions}{Logical. If TRUE only partitions above a particular
threshold value are selected.}

\item{cut_level}{Which selecting threshold will be used to cut the mean
models. Default is set to "spec_sens" but any \pkg{dismo} threshold (see
function \code{\link[dismo]{threshold}}) can be used: "kappa", "no_omission",
 "prevalence", "equal_sens_spec", "sensitivity".}

\item{scale_models}{Logical. Whether input models should be scaled between 0
and 1.}

\item{select_par}{Which performance statistic should be used to select the
partitions. Defaults to NULL but either \code{"AUC"} or \code{"TSS"} can be
used.}

\item{select_par_val}{Threshold to select models from TSS values.}

\item{consensus_level}{Which proportion of models will be kept when creating
\code{bin_consensus} (binary).}

\item{models_dir}{Character. Folder path where the input files are located.}

\item{final_dir}{Character. Name of the folder to save the output files.
A subfolder will be created.}

\item{proj_dir}{Character. The name of the subfolder with the projection.
Defaults to "present" but can be set according to the other projections (i.e.
to execute the function in projected models).}

\item{which_models}{Which \code{final_model} will be used? Currently it can
be:
\describe{
  \item{\code{raw_mean}}{Continuous model generated by the mean of the raw
  models (scale from 0 to 1)}
  \item{\code{raw_mean_th}}{Cuts the \code{raw_mean} by the mean of the
  thresholds that maximize the selected evaluation metric (e.g. TSS
  (\code{spec_sens}) to make a binary model.}
   \item{\code{raw_mean_cut}}{Recovers \code{raw_mean} values above the mean
   threshold that maximizes the selected evaluation metric (e.g. TSS
   (\code{spec_sens}) or other \pkg{dismo} thresholds). Generates a
   continuous model.}
  \item{\code{bin_mean}}{The mean of the selected binary models. Generates a
  model in a discrete scale.}
  \item{\code{bin_consensus}}{The binary consensus from \code{bin_mean}.
  Parameter \code{consensus_level} must be defined, 0.5 means a majority
  consensus}
  \item{\code{cut_mean}}{The mean of the selected cut models. Values below
  the thresholds are down-weighted by zeros.}
}}

\item{uncertainty}{Whether an uncertainty map, measured as range (max-min)
should be calculated.}

\item{write_final}{Logical. If \code{TRUE}, writes png files of the final
models.}

\item{...}{Other parameters from \code{\link[raster]{writeRaster}}}
}
\value{
A set of ecological niche models (.tif files) and figures (optional)
written in the \code{final_dir} subfolder.
}
\description{
This function reads the models generated either by \code{\link{do_any}} or
\code{\link{do_many}} (i.e. one model per partition per algorithm) and
summarizes them into a final model for each species-algorithm combination.
These final models may be created from the raw uncut models, the binary
models or the "cut" models (see \code{\link{do_any}}). They may use all
partitions or select only partitions with performance metrics above a defined
value. The mean can be calculated as arithmetic mean or weighted by a
performance statistic (AUC or TSS). A majority rule can be applied to binary
models and uncertainty taken as ranges between paritition may also be
calculated.
}
\examples{
# run setup_sdmdata
sp <- names(coordenadas)[1]
sp_coord <- coordenadas[[1]]
sp_setup <- setup_sdmdata(species_name = sp, occurrences = sp_coord,
example_vars)

# run do_any
sp_bioclim <- do_any(species_name = sp,
                     predictors = example_vars,
                     algo = "bioclim")

# run final_model
sp_final <- final_model(species_name = sp,
                        algorithms = "bioclim",
                        select_partitions = TRUE,
                        select_par = "TSS",
                        select_par_val = 0,
                        which_models = c("bin_consensus"),
                        consensus_level = 0.5,
                        overwrite = TRUE)

}
\seealso{
\code{\link[dismo]{threshold}}

\code{\link[raster]{writeRaster}}
}
