% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_model.R
\name{final_model}
\alias{final_model}
\title{Joins ENM from several partitions, creating a model per algorithm.}
\usage{
final_model(species.name, algorithms = NULL, weight.par = NULL,
  select.partitions = TRUE, TSS.value = 0.7, threshold = c("spec_sens"),
  consensus.level = 0.5, models.dir = "./models",
  final.dir = "final_models")
}
\arguments{
\item{species.name}{A character string with the species name}

\item{algorithms}{Which algorithms will be processed. If no name is given it
will process all algorithms present in the evaluation files}

\item{weight.par}{Which performance statistic should be used to weight the
partitions defaults to NULL but either \code{c("AUC", "TSS")} can be used.}

\item{select.partitions}{TRUE ou FALSE}

\item{TSS.value}{Threshold to select models from TSS values}

\item{threshold}{Which selecting threshold will be used to cut the mean
models in final_model_3 approach (see vignettes), it
defaults to "spec_sens" but any dismo threshold
can be used: "kappa", "no_omission", "prevalence",
"equal_sens_spec", "sensitivity".}

\item{consensus.level}{Which proportion of models will be kept when creating
\code{final_model_8} (binary)}

\item{models.dir}{Character. Folder path where the input files are located}

\item{final.dir}{Character. Name of the folder to save the output files.
A subfolder will be created.}
}
\value{
A set of ecological niche models and figures (optional) written in the
         \code{final.dir} subfolder
}
\description{
This function reads the output from dismo.mod and creates a model per species
}
