% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_model.R
\name{finalModel}
\alias{finalModel}
\title{Gera um modelo final por algortimo}
\usage{
finalModel(sp, select.partitions = T, algoritmos = c("maxent", "rf", "svm"),
  threshold = c("spec_sens"), TSS.value = 0.7, models.dir = "./models",
  final.dir = "final_models")
}
\arguments{
\item{sp}{Um nome de espécie}

\item{select.partitions}{TRUE ou FALSE}

\item{threshold}{"spec_sens" por padrão}

\item{TSS.value}{0.7 por padrão}

\item{models.dir}{Path do diretório onde estão os modelos}

\item{final.dir}{Em que subdiretório escrever os modelos finais}

\item{algortimos}{Algortimos selecionados. Maxent, RF e SVM por padrão}
}
\value{
Uma string com a data e hora do término da execução
}
\description{
Gera um modelo final por algortimo
}
