% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_filt.R
\name{geo_filt}
\alias{geo_filt}
\title{Excludes occurrence points within a given geographic distance}
\usage{
geo_filt(occurrences, lon = "lon", lat = "lat", min_distance = 10)
}
\arguments{
\item{occurrences}{A data frame with occurrence data. Data must have at least
 columns with latitude and longitude values of species occurrences.
See \code{\link{coordenadas}} as an example}

\item{lon}{The name of the longitude column. Defaults to "lon"}

\item{lat}{The name of the latitude column. Defaults to "lat"}

\item{min_distance}{Numeric. Minimum distance between points in the unit of
the predictor raster. This value will be used to create a RaserStack object
with a \code{min_distance} resolution, from which the redundant occurences
will be excluded}
}
\value{
Data frame of the occurrence points thinned to have the minimum
distance indicated in \code{min_distance} between them
}
\description{
Excludes occurrence points within a given geographic distance
}
\references{
Varela, S., Anderson, R. P., García-Valdés, R., &
Fernández-González, F. (2014). Environmental filters reduce the effects of
sampling bias and improve predictions of ecological niche models.
Ecography, 37(11), 1084-1091. doi:10.1111/j.1600-0587.2013.00441.x
}
\seealso{
\code{\link[dismo]{gridSample}}

\pkg{spThin}
}
