% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_filt.R
\name{geo_filt}
\alias{geo_filt}
\title{Excludes occurrence points within a geographic distance}
\usage{
geo_filt(coordinates, min_distance = 10)
}
\arguments{
\item{coordinates}{data.frame. Table with the species' data. It should contain only two columns: lon and lat, in that order.}

\item{min_distance}{numeric. Minimum distance (in Km) between points.}
}
\value{
Table of occurrence points with minimum distance indicated in min.distance
}
\description{
Excludes occurrence points within a geographic distance
}
\references{
Varela, S., Anderson, R. P., García-Valdés, R., &
Fernández-González, F. (2014). Environmental filters reduce the effects of
sampling bias and improve predictions of ecological niche models.
Ecography, 37(11), 1084-1091. doi:10.1111/j.1600-0587.2013.00441.x
}
\seealso{
\code{\link[dismo]{gridSample}}
}
\author{
Diogo S. B. Rocha
}
