% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mascara}
\alias{mascara}
\title{Mask of the Brazilian Atlantic Forest, based on IBGE}
\format{A SpatialPolygonsDataFrame of the Brazilian Atlantic Forest}
\usage{
mascara
}
\description{
Mask of the Brazilian Atlantic Forest, based on IBGE
}
\keyword{datasets}
