% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale_layer}
\alias{rescale_layer}
\title{Rescaling rasters to values between 0 and 1.}
\usage{
rescale_layer(layers)
}
\arguments{
\item{layers}{A RasterStack or Rasterlayer of predictor variables to scale}
}
\value{
A RasterStack or Rasterlayer with values between 0 and 1.
}
\description{
This function rescales rasters to values between 0 and 1.
}
\examples{
rescale_layer(example_vars)
}
\author{
Diogo S. B. Rocha
}
