% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_variables.R
\name{select_variables}
\alias{select_variables}
\title{Helper function to select environmental variables from a stack}
\usage{
select_variables(species_name = species_name, predictors = predictors,
  models_dir = "./models_dir", buffer = NULL, cutoff = 0.8, ...)
}
\arguments{
\item{species_name}{A character string with the species name}

\item{predictors}{Stack of environmental variables}

\item{models_dir}{Character name of output directory where the
buffer created with create_buffer() is}

\item{buffer}{Shapefile specified by user or use NULL to use the output from create_buffer()}

\item{cutoff}{Cutoff value of correlation between variables to exclude environmental layer. Default is to exclude environmental variables with correlation > 0.8.}

\item{...}{parameters from create_buffer()}
}
\value{
A raster stack of independent environmental variables based on a
 specific cutoff
}
\description{
This function takes a stack of environmental variables and the calibration area. The function calculates Pearson correlations between all environmental variables and returns a new stack with uncorrelated variables based on a cutoff.
}
\seealso{
\code{\link[ModelR]{create_buffer}}
}
\author{
Andrea Sánchez-Tapia and Sara Mortara
}
