% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_sdmdata.R
\name{setup_sdmdata}
\alias{setup_sdmdata}
\title{Prepares the dataset to perform ENM}
\usage{
setup_sdmdata(species_name = species_name, coordinates = coordinates,
  real_absences = NULL, lon = "lon", lat = "lat", buffer = FALSE,
  seed = 512, predictors = predictors, clean_dupl = T, clean_nas = F,
  geo_filt = F, geo_filt_dist, models.dir = models.dir, plot_sdmdata = T,
  n.back = 1000, bootstrap = F, boot_proportion = 0.8, boot_n = 10,
  crossvalidation = F, cv_partitions = cv_partitions, cv_n = 10)
}
\arguments{
\item{species_name}{A character string with the species name}

\item{coordinates}{A data frame with occurrence data}

\item{real_absences}{User defined absence points}

\item{lon}{the name of the longitude column. defaults to "lon"}

\item{lat}{the name of the latitude column. defaults to "lat"}

\item{buffer}{Defines if a buffer will be used to sample pseudo-absences
(F, "mean", "median", "max")}

\item{seed}{For reproducibility purposes}

\item{predictors}{A RasterStack of predictor variables}

\item{clean_dupl}{Logical, delete duplicate occurrence points? defaults to
TRUE}

\item{clean_nas}{Logical, delete occurrence points with no environmental
information? Defaults to FALSE and can take a while for large datasets}

\item{geo_filt}{Logical, delete occurrence that are too close?}

\item{geo_filt_dist}{The distance of the geographic filter (in kilometers)}

\item{models.dir}{Folder path to save the output files}

\item{plot_sdmdata}{Logical, whether png files will be written}

\item{n.back}{Number of pseudoabsence points}

\item{bootstrap}{Logical, perform bootstrap}

\item{boot_proportion}{Numerical 0 to 1, proportion of points to be sampled
for bootstrap}

\item{boot_n}{How many bootstrap runs}

\item{crossvalidation}{Logical, perform k-fold crossvalidation}

\item{cv_partitions}{Number of partitions in the crossvalidation}

\item{cv_n}{How many crossvalidation runs}
}
\value{
A dataframe called sdmdata with the groups for each run
(in columns called cv.1, cv.2 or boot.1, boot.2), a presence/absence vector,
the geographical coordinates, of the occurrence and pseudoabsence points, and
the associated environmental variables.
}
\description{
Prepares the dataset to perform ENM
}
\seealso{
\code{\link[dismo]{gridSample}}
}
\author{
Andrea Sánchez-Tapia
}
