% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variaveis_preditoras}
\alias{variaveis_preditoras}
\title{Predictor variables}
\format{A RasterStack with 6 predictor variables issued from a PCA}
\usage{
variaveis_preditoras
}
\description{
Predictor variables
}
\keyword{datasets}
